+++ feature/navigation-adaptive/build.gradle.kts
@@ -47,6 +47,7 @@ dependencies {
 
     // Adaptive dependencies
     implementation(libs.androidx.material3.adaptive)
+    implementation(libs.androidx.material3.adaptive.layout)
     implementation(libs.androidx.material3.adaptive.navigation.suite.android)
 
     // Navigation dependencies
