+++ gradle/libs.versions.toml
@@ -8,7 +8,7 @@ espressoCore = "3.7.0"
 lifecycleRuntimeKtx = "2.9.4"
 activityCompose = "1.11.0"
 composeBom = "2025.09.00"
-material3Adaptive = "1.1.0"
+material3Adaptive = "1.2.0-beta02"
 navigationCompose = "2.9.4"
 kotlinSerialization = "1.9.0"
