+++ settings.gradle.kts
@@ -24,7 +24,6 @@ include(":app")
 include(":core:config")
 include(":core:route")
 include(":core:ui")
-include(":feature:navigation")
 include(":feature:navigation-base")
 include(":feature:navigation-adaptive")
 include(":feature:navigation-not-adaptive")
